requirements.txt
logging
RPi.GPIO